requirements.txt
matplotlib
django
djangorestframework